requirements.txt
discord.py[voice]
python-dotenv
asyncpg
